Refactor/rename Emails to use Smartyer templates for email content
FossilOrigin-Name: f10ad977cd1e7a392a9d12e2942dcbe32a31364330f62ab5037d68326951420f

+++ src/www/admin/password.php
@@ -32,7 +32,7 @@ $csrf_key = 'recover_password';
 $new = qg('new') !== null;
 
 $form->runIf('recover', function () use ($session) {
-	$session->recoverPasswordSend((int) f('id'));
+	$session->recoverPasswordSend(f('id'));
 }, $csrf_key, '!password.php?sent' . ($new ? '&new' : ''));
 
 $sent = !$form->hasErrors() && null !== qg('sent');
